// types/circuit.ts
export interface Circuit {
  id: string;
  slug: string;
  name: string;
  country: string;
  flag: string;
  location: string;
  length: string;
  laps: number;
  raceDistance: string;
  firstGrandPrix: number;
  lapRecord: string;
  corners: number;
  drsZones: number;
  image: string;
  description: string;
}